ASoC: SOF: ipc4-topology: set copier output format for external module
The copier output pin 0 format is set with module instance
initialization, format for additional copier output pin
should be set before the pin is used.

If an external module is connected to additional copier output
pin, the copier output pin format should be set according to
the corresponding sink pin format of the external module.

Signed-off-by: Chao Song <chao.song@linux.intel.com>