ASoC: SOF: ipc4-topology: add base module config extension structure
Legacy modules use the base module config containing the audio format
for the source/sink pin 0 for initialization. But some processing modules
with multiple source/sink pins require the audio formats to be sent with
the init instance IPC payload.

Introduce a new struct sof_ipc4_base_module_cfg_ext which contains the
pin formats for the source and sink pins. This will be appended to the
init instance IPC payload for modules that require it.
Signed-off-by: Chao Song <chao.song@linux.intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>